Massive Edit
Deleted multiple files
Deleted multiple unused imports
Tried to add comments to make things more readable
Swapped out the vgg in train.py and loss.py since it was unused
The code needs to be tested that it still runs with these changes
The models especially need to be commented
I believe function.py is unused but held off on deleting it in case it has some weird interaction with math library
Do not merge with main yet as needs to test that it still works
Transformer.py needs to be commented
unet.py is only used in one place and I believe it is a testing area. Consider moving test aspects to another branch besides main.